requirements.txt
-i https://pypi.org/simple
--extra-index-url http://pypi.19gale.ai:9999/simple
--trusted-host pypi.19gale.ai
absl-py==0.12.0
aniso8601==9.0.1
appnope==0.1.2 ; sys_platform == 'darwin'
astunparse==1.6.3
attrs==20.3.0
backcall==0.2.0
cachetools==4.2.1
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
dataclasses-json==0.5.2
decorator==5.0.7
flask-httpauth==3.3.0
flask-restful==0.3.8
flask-sqlalchemy==2.4.1
flask==1.1.1
flatbuffers==1.12
gast==0.3.3
google-auth-oauthlib==0.4.4
google-auth==1.29.0
google-pasta==0.2.0
graphviz==0.16
greenlet==1.0.0 ; python_version >= '3'
grpcio==1.32.0
h5py==2.10.0
idna==2.8
importlib-metadata==3.10.1 ; python_version < '3.8'
ipykernel==5.5.3
ipython-genutils==0.2.0
ipython==7.22.0
itsdangerous==1.1.0
jedi==0.18.0
jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.2.0
jupyter-core==4.7.1
keras-preprocessing==1.1.2
keras==2.4.3
loguru==0.4.1
markdown==3.3.4
markupsafe==1.1.1
marshmallow-enum==1.5.1
marshmallow==3.11.1
mypy-extensions==0.4.3
nbformat==5.1.3
nest-asyncio==1.5.1
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.1.2
parso==0.8.2
pexpect==4.8.0 ; sys_platform != 'win32'
pickleshare==0.7.5
plotly==4.14.3
prompt-toolkit==3.0.18
protobuf==3.15.8
psycopg2-binary==2.8.3
ptyprocess==0.7.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pydot==1.4.2
pygments==2.8.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.10.3
pytz==2019.3
pyyaml==5.4.1
pyzmq==22.0.3
redis==3.5.3
requests-oauthlib==1.3.0
requests==2.22.0
retrying==1.3.3
rsa==4.7.2 ; python_version >= '3.6'
scikit-learn==0.24.1
scipy==1.6.2
six==1.15.0
sqlalchemy==1.4.9
stringcase==1.2.0
tensorboard-plugin-wit==1.8.0
tensorboard==2.4.1
tensorflow-estimator==2.4.0
tensorflow==2.4.1
termcolor==1.1.0
threadpoolctl==2.1.0
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3 ; python_version < '3.8'
typing-inspect==0.6.0
urllib3==1.25.11
wcwidth==0.2.5
werkzeug==1.0.1
wheel==0.36.2 ; python_version >= '3'
wrapt==1.12.1
zipp==3.4.1
